+++ plugins/input/httpserver/input_http_server.go
@@ -24,6 +24,7 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/alibaba/ilogtail/helper"
 	"github.com/alibaba/ilogtail/helper/decoder"
 	"github.com/alibaba/ilogtail/helper/decoder/common"
 	"github.com/alibaba/ilogtail/pkg/logger"
@@ -36,6 +37,8 @@ const (
 	v2
 )
 
+const name = "service_http_server"
+
 // ServiceHTTP ...
 type ServiceHTTP struct {
 	context     pipeline.Context
@@ -47,15 +50,20 @@ type ServiceHTTP struct {
 	collectorV2 pipeline.PipelineCollector
 	version     int8
 	paramCount  int
+	dumper      *helper.Dumper
 
+	DumpDataKeepFiles  int
+	DumpData           bool // would dump the received data to a local file, which is only used to valid data by the developers.
 	Format             string
 	Address            string
+	Path               string
 	ReadTimeoutSec     int
 	ShutdownTimeoutSec int
 	MaxBodySize        int64
 	UnlinkUnixSock     bool
 	FieldsExtend       bool
 	DisableUncompress  bool
+	Tags               map[string]string // todo for v2
 
 	// params below works only for version v2
 	QueryParams       []string
@@ -72,17 +80,27 @@ func (s *ServiceHTTP) Init(context pipeline.Context) (int, error) {
 		return 0, err
 	}
 
-	switch s.Format {
-	case common.ProtocolOTLPLogV1:
-		s.Address += "/v1/logs"
-	case common.ProtocolOTLPMetricV1:
-		s.Address += "/v1/metrics"
-	case common.ProtocolOTLPTraceV1:
-		s.Address += "/v1/traces"
+	if s.Path == "" {
+		switch s.Format {
+		case common.ProtocolOTLPLogV1:
+			s.Path = "/v1/logs"
+		case common.ProtocolOTLPMetricV1:
+			s.Path = "/v1/metrics"
+		case common.ProtocolOTLPTraceV1:
+			s.Path = "/v1/traces"
+		case common.ProtocolPyroscope:
+			s.Path = "/ingest"
+		}
 	}
+	s.Address += s.Path
+	logger.Infof(context.GetRuntimeContext(), "addr", s.Address, "format", s.Format)
 
 	s.paramCount = len(s.QueryParams) + len(s.HeaderParams)
 
+	if s.DumpData {
+		s.dumper = helper.NewDumper(strings.Join([]string{name, context.GetProject(), context.GetConfigName()}, "-"), s.DumpDataKeepFiles)
+		s.dumper.Init()
+	}
 	return 0, nil
 }
 
@@ -103,8 +121,7 @@ func (s *ServiceHTTP) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	data, statusCode, err := s.decoder.ParseRequest(w, r, s.MaxBodySize)
-	logger.Debugf(s.context.GetRuntimeContext(), "request [method] %v; [header] %v; [url] %v; [body] %v", r.Method, r.Header, r.URL, string(data))
-
+	logger.Debugf(s.context.GetRuntimeContext(), "request [method] %v; [header] %v; [url] %v; [body len] %d", r.Method, r.Header, r.URL, len(data))
 	switch statusCode {
 	case http.StatusBadRequest:
 		BadRequest(w)
@@ -114,16 +131,24 @@ func (s *ServiceHTTP) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		InternalServerError(w)
 	case http.StatusMethodNotAllowed:
 		MethodNotAllowed(w)
-
 	}
 	if err != nil {
 		logger.Warning(s.context.GetRuntimeContext(), "READ_BODY_FAIL_ALARM", "read body failed", err, "request", r.URL.String())
 		return
 	}
 
+	if s.dumper != nil {
+		s.dumper.InputChannel() <- &helper.DumpData{
+			Req: helper.DumpDataReq{
+				Body:   data,
+				URL:    r.URL.String(),
+				Header: r.Header,
+			},
+		}
+	}
 	switch s.version {
 	case v1:
-		logs, err := s.decoder.Decode(data, r)
+		logs, err := s.decoder.Decode(data, r, s.Tags)
 		if err != nil {
 			logger.Warning(s.context.GetRuntimeContext(), "DECODE_BODY_FAIL_ALARM", "decode body failed", err, "request", r.URL.String())
 			BadRequest(w)
@@ -147,13 +172,15 @@ func (s *ServiceHTTP) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		s.collectorV2.CollectList(groups...)
 	}
 
-	if s.Format == "sls" {
+	switch s.Format {
+	case common.ProtocolSLS:
 		w.Header().Set("x-log-requestid", "1234567890abcde")
 		w.WriteHeader(http.StatusOK)
-	} else {
+	case common.ProtocolPyroscope:
+		// do nothing
+	default:
 		w.WriteHeader(http.StatusNoContent)
 	}
-
 }
 
 func TooLarge(res http.ResponseWriter) {
@@ -190,7 +217,6 @@ func (s *ServiceHTTP) Start(c pipeline.Collector) error {
 func (s *ServiceHTTP) StartService(context pipeline.PipelineContext) error {
 	s.collectorV2 = context.Collector()
 	s.version = v2
-
 	return s.start()
 }
 
@@ -236,6 +262,9 @@ func (s *ServiceHTTP) start() error {
 		logger.Info(s.context.GetRuntimeContext(), "http server shutdown", s.Address)
 		s.wg.Done()
 	}()
+	if s.dumper != nil {
+		s.dumper.Start()
+	}
 	return nil
 }
 
@@ -277,16 +306,21 @@ func (s *ServiceHTTP) Stop() error {
 		logger.Info(s.context.GetRuntimeContext(), "http server stop", s.Address)
 		s.wg.Wait()
 	}
+	if s.dumper != nil {
+		s.dumper.Close()
+	}
 	return nil
 }
 
 func init() {
-	pipeline.ServiceInputs["service_http_server"] = func() pipeline.ServiceInput {
+	pipeline.ServiceInputs[name] = func() pipeline.ServiceInput {
 		return &ServiceHTTP{
 			ReadTimeoutSec:     10,
 			ShutdownTimeoutSec: 5,
 			MaxBodySize:        64 * 1024 * 1024,
 			UnlinkUnixSock:     true,
+			DumpDataKeepFiles:  5,
+			Tags:               map[string]string{},
 		}
 	}
 }