+++ plugins/input/httpserver/input_http_server_test.go
@@ -29,7 +29,6 @@ import (
 	pluginmanager "github.com/alibaba/ilogtail/pluginmanager"
 
 	"github.com/stretchr/testify/require"
-
 	"gotest.tools/assert"
 )