+++ CHANGELOG.md
@@ -55,3 +55,5 @@ your changes, such as:
 - [public] [both] [fixed] fix plugin version in logs
 - [public] [both] [updated] flusher http support variable config in request header
 - [public] [both] [fixed] fix memory leak in container list maintainance introduced in v1.2.1
+- [public] [both] [added] support pyroscope input datasource
+- [public] [both] [added] support config plugins to included in build, plugins.yml for builtin plugins, external_plugins.yml for external plugins